requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.9

# ML/NLP - use flexible versions for Windows compatibility
torch>=2.2.0
transformers==4.37.0
sentence-transformers==2.3.0
accelerate>=0.26.0

# Vector DB & RAG
chromadb==0.4.22
langchain>=0.1.0
langchain-community>=0.0.17
langchain-openai>=0.0.5

# LLM
openai==1.12.0

# Data & Utils - use flexible versions
numpy>=1.24.0
pandas>=2.0.0
python-dotenv==1.0.0
redis>=5.0.0
httpx>=0.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
